raspberry/pyproject.toml
[project]
name = "raspberry"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "serial (>=0.0.97,<0.0.98)",
    "aioserial (>=1.3.1,<2.0.0)",
    "aiomqtt (>=2.4.0,<3.0.0)",
    "pymongo (>=4.15.4,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
